from and to: allow http(s) comparison
Enable the use of http(s) URLs to identify lock files.
Files are already loaded via generic stream wrappers but not all stream wrappers,
http(s) in particular[0], support the used checking ("stat") for the existence of the
file.

Would be happy to supply testing and refactoring beyond this naive solution once the
maintainer outlines some paradigms.

Sample use:
./composer-lock-diff --from https://github.com/wikimedia/mediawiki-vendor/raw/REL1_29/composer.lock --to https://github.com/wikimedia/mediawiki-vendor/raw/REL1_30/composer.lock

Please note that, for github repositories, you need to provide URLs of the "raw"
version of files to ensure a comparison can be performed.

[0] https://secure.php.net/manual/en/wrappers.http.php#refsect1-wrappers.http-options